There's a bit of special logic needed to parse the `saveProjectReviewCLISurveyResponses` API inputs and add in the extra context that's missing. Previously this branch was triggering all of that logic deeper in the stack in an action based on whether an individual response had a 'values' attribute. Since the logic is primarily about handling that particular API endpoints inputs I've moved the logic back into the resolver layer with a custom resolver for that endpoint that calls a helper to convert the sparse named responses received by that endpoint into "full" responses with all of the context they need to be saved.
I've also moved some of the authentication checking closer to the API.
For example, checking to make sure that players don't submit responses
for others is now done in the resolver, not the action. This means we
could use this action in an admin interface where we might want to
submit responses for another user, or in a script where there's no
concept of a "current user" at all.

I also noticed that we were using `findActiveProjectReviewSurvey` to get
the "active" project review survey for a project. I suspect this was a
holdover from a time when the project data model supported a project
that spanned multiple cycles. Since that's not the case anymore I
switched to a simple thinky `Survey.get()` to get the survey by id and
removed the `findActiveProjectReviewSurvey` method which is no longer
used anywhere.