Adding projectCreated worker.
This worker now handles creating the channels and sending th welcome
messages. Since this is now asynchronous I've added a delay to message
that goes to the chapter channel so that all of the rooms will be
created before the list is posted there. This is needed because if the
channels are not created at the time the message is posted than their
names will not be clickable (even after they do get created, unlike
slack).