Amendments reflecting current chat-integration conditions.
To avoid API errors in the execution of “npm run workers”, one would already need the 2 Slack-related entries in the .env.development file that were shown later on the page. This would require getting invited by a teammate to the Slack dev team. But current P4 members are not aware of such a team, are not members of it, and cannot give invitations. Therefore it is appropriate to warn about the error message and delete the entirety of the section on Slack integration.